pyproject.toml
[tool.poetry]
name = "part2"
version = "0.1.0"
description = ""
authors = ["Tungnt24 <nguyentutung24@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
celery = "^5.0.3"
pyTelegramBotAPI = "^3.7.4"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
